kauai/src/dlgwin.cpp: use uintptr_t cast for pointer alignment check
To avoid pointer truncation on 64-bit Windows during the pointer
alignment check, cast the pointer to uintptr_t instead of int32_t.

+++ kauai/src/dlgwin.cpp
@@ -111,7 +111,7 @@ bool DLG::_FInit(void)
     for (csit = dtm.cdit, idit = 0; csit > 0; csit--)
     {
         // align to dword
-        if ((int32_t)psw & 2)
+        if ((uintptr_t)psw & 2)
             psw++;
 
         // get and skip the ditm
